style: detail/basic
