+++ frontend/providers/devbox/message/en.json
@@ -183,7 +183,7 @@
   "none": "None",
   "none_template_version": "No versions of this Runtime are currently available, rendering it unusable. To restore functionality, please add a new version in the project details page.",
   "not_allow_standalone_use": "Not allowed to use standalone",
-  "one_click_config": "One-Click Setup",
+  "one_click_config": "SSH Connection Setup",
   "open_link": "Open link",
   "open_vscode": "VS Code",
   "opening_ide": "Opening IDE...",
@@ -238,7 +238,7 @@
   "shutdown": "Shutdown",
   "skip": "Skip",
   "ssh_config": "SSH Configuration",
-  "ssh_connect_info": "SSH Connection String",
+  "ssh_connect_info": "SSH Connection",
   "start": "Start",
   "start_devbox": "Start",
   "start_error": "Start error",
